requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
billiard==3.3.0.23
Django==3.0.6
django-celery-beat==2.0.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
